refactor: match 테스트용 더미데이터 생성 서비스 패키지 분리
